build framework mmdebstrap - fetch the newest keyring pkgs

+++ lib/functions/rootfs/distro-specific.sh
@@ -92,40 +92,59 @@ function fetch_distro_keyring() {
 			if [ -e "${CACHEDIR}/debian-archive-keyring.gpg" ]; then
 				display_alert "fetch_distro_keyring($release)" "cache found, skipping" "info"
 			else
-				BASEURI='https://deb.debian.org/debian/pool/main/d/debian-archive-keyring/'
-				#FIXME: write something to retrieve newest
-				KEYRING_DEB='debian-archive-keyring_2025.1_all.deb'
-				curl -fLOJ --output-dir "${CACHEDIR}" "${BASEURI}/${KEYRING_DEB}" || \
-					exit_with_error "fetch_distro_keyring failed" "unable to download ${BASEURI}/${KEYRING_DEB}"
+			# FIXME: should this be a loop? might allow more parameterization and shorter lines to read.
+				PKG_URL=$(curl --compressed -Ls 'https://packages.debian.org/sid/all/debian-archive-keyring/download' | \
+					grep -oP 'http://(deb|ftp)\.debian\.org/debian/pool/main/d/debian-archive-keyring/debian-archive-keyring_[0-9.]*_[a-zA-Z0-9]*\.deb')
+				[[ -z "${PKG_URL}" ]] && exit_with_error "fetch_distro_keyring failed" "unable to find newest debian-archive-keyring package"
+				curl -fLOJ --output-dir "${CACHEDIR}" "${PKG_URL}" || \
+					exit_with_error "fetch_distro_keyring failed" "unable to download ${PKG_URL}"
+				KEYRING_DEB=$(basename "${PKG_URL}")
 				dpkg-deb -x "${CACHEDIR}/${KEYRING_DEB}" "${CACHEDIR}" || \
-					exit_with_error "fetch_distro_keyring" "dpkg-deb -x ${KEYRING_DEB} failed"
-				# yes, for 2025.1, the canonical name is .pgp, but our tools expect .gpg.
-				# the package contains the .pgp and a .gpg symlink to it.
-				cp -l "${CACHEDIR}/usr/share/keyrings/debian-archive-keyring.pgp" "${CACHEDIR}/debian-archive-keyring.gpg"
+					exit_with_error "fetch_distro_keyring" "dpkg-deb -x ${CACHEDIR}/${KEYRING_DEB} failed"
+				if [[ -e "${CACHEDIR}/usr/share/keyrings/debian-archive-keyring.pgp" ]]; then
+					# yes, for 2025.1, the canonical name is .pgp, but our tools expect .gpg.
+					# the package contains the .pgp and a .gpg symlink to it.
+					cp -l "${CACHEDIR}/usr/share/keyrings/debian-archive-keyring.pgp" "${CACHEDIR}/debian-archive-keyring.gpg"
+				elif [[ -e "${CACHEDIR}/usr/share/keyrings/debian-archive-keyring.gpg" ]]; then
+					cp -l "${CACHEDIR}/usr/share/keyrings/debian-archive-keyring.gpg" "${CACHEDIR}/debian-archive-keyring.gpg"
+				else
+					exit_with_error "fetch_distro_keyring" "unable to find debian-archive-keyring.gpg"
+				fi
 				display_alert "fetch_distro_keyring($release)" "extracted" "info"
 
-				BASEURI='https://deb.debian.org/debian/pool/main/d/debian-ports-archive-keyring/'
-				#FIXME: write something to retrieve newest
-				KEYRING_DEB='debian-ports-archive-keyring_2025.04.05_all.deb'
-				curl -fLOJ --output-dir "${CACHEDIR}" "${BASEURI}/${KEYRING_DEB}" || \
-					exit_with_error "fetch_distro_keyring failed" "unable to download ${BASEURI}/${KEYRING_DEB}"
+				PKG_URL=$(curl --compressed -Ls 'https://packages.debian.org/sid/all/debian-ports-archive-keyring/download' | \
+					grep -oP 'http://(deb|ftp)\.debian\.org/debian/pool/main/d/debian-ports-archive-keyring/debian-ports-archive-keyring_[0-9.]*_[a-zA-Z0-9]*\.deb')
+				[[ -z "${PKG_URL}" ]] && exit_with_error "fetch_distro_keyring failed" "unable to find newest debian-ports-archive-keyring package"
+				curl -fLOJ --output-dir "${CACHEDIR}" "${PKG_URL}" || \
+					exit_with_error "fetch_distro_keyring failed" "unable to download ${PKG_URL}"
+				KEYRING_DEB=$(basename "${PKG_URL}")
 				dpkg-deb -x "${CACHEDIR}/${KEYRING_DEB}" "${CACHEDIR}" || \
-					exit_with_error "fetch_distro_keyring" "dpkg-deb -x ${KEYRING_DEB} failed"
-				# see above comment about .pgp vs .gpg
-				cp -l "${CACHEDIR}/usr/share/keyrings/debian-ports-archive-keyring.pgp" "${CACHEDIR}/debian-ports-archive-keyring.gpg"
+					exit_with_error "fetch_distro_keyring" "dpkg-deb -x ${CACHEDIR}/${KEYRING_DEB} failed"
+				if [[ -e "${CACHEDIR}/usr/share/keyrings/debian-ports-archive-keyring.pgp" ]]; then
+					# see above comment re .pgp vs .gpg
+					cp -l "${CACHEDIR}/usr/share/keyrings/debian-ports-archive-keyring.pgp" "${CACHEDIR}/debian-ports-archive-keyring.gpg"
+				elif [[ -e "${CACHEDIR}/usr/share/keyrings/debian-ports-archive-keyring.gpg" ]]; then
+					cp -l "${CACHEDIR}/usr/share/keyrings/debian-ports-archive-keyring.gpg" "${CACHEDIR}/debian-ports-archive-keyring.gpg"
+				else
+					exit_with_error "fetch_distro_keyring" "unable to find debian-ports-archive-keyring.gpg"
+				fi
 			fi
 			;;
 		ubuntu)
 			if [ -e "${CACHEDIR}/ubuntu-archive-keyring.gpg" ]; then
 				display_alert "fetch_distro_keyring($release)" "cache found, skipping" "info"
 			else
-				BASEURI='https://archive.ubuntu.com/ubuntu/pool/main/u/ubuntu-keyring/'
-				#FIXME: write something to retrieve newest
-				KEYRING_DEB='ubuntu-keyring_2023.11.28.1_all.deb'
-				curl -fLOJ --output-dir "${CACHEDIR}" "${BASEURI}/${KEYRING_DEB}" || \
-					exit_with_error "fetch_distro_keyring failed" "unable to download ${BASEURI}/${KEYRING_DEB}"
+				NEWEST_SUITE=$(curl --compressed -sv https://changelogs.ubuntu.com/meta-release | grep 'Dist:'|tail -n 1 | awk '{print $NF}')
+				PKG_URL=$(curl --compressed -Ls "https://packages.ubuntu.com/${NEWEST_SUITE}/all/ubuntu-keyring/download" | \
+					grep -oP 'http://\S+\.deb' |grep archive.ubuntu.com|tail -n 1)
+				[[ -z "${PKG_URL}" ]] && exit_with_error "fetch_distro_keyring failed" "unable to find newest ubuntu-keyring package"
+				# ubuntu gives a long list of regional mirrors, we want as generic as possible
+				PKG_URL=$(echo "${PKG_URL}" | sed -E 's/[a-z0-9]+\.archive/archive/')
+				curl -fLOJ --output-dir "${CACHEDIR}" "${PKG_URL}" || \
+					exit_with_error "fetch_distro_keyring failed" "unable to download ${PKG_URL}"
+				KEYRING_DEB=$(basename "${PKG_URL}")
 				dpkg-deb -x "${CACHEDIR}/${KEYRING_DEB}" "${CACHEDIR}" || \
-					exit_with_error "fetch_distro_keyring" "dpkg-deb -x ${KEYRING_DEB} failed"
+					exit_with_error "fetch_distro_keyring" "dpkg-deb -x ${CACHEDIR}/${KEYRING_DEB} failed"
 				cp -l "${CACHEDIR}/usr/share/keyrings/ubuntu-archive-keyring.gpg" "${CACHEDIR}/"
 			fi
 			debootstrap_arguments+=("--keyring=/usr/share/keyrings/ubuntu-archive-keyring.gpg")
